s2a: Move IntegrationTest to io.grpc.s2a
It uses a io.grpc.s2a class, so shouldn't be in internal.handler
